refactor: use presentational components
