+++ packages/demo/frontend/src/App.tsx
@@ -1,8 +1,9 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import Terminal from './components/Terminal'
 import Home from './components/home/Home'
-import Earn from './components/Earn'
 import { PrivyProvider } from './providers/PrivyProvider'
+import { EarnWithPrivyServerWallet } from './components/EarnWithPrivyServerWallet'
+import { ActivityLogProvider } from './contexts/ActivityLogContext'
 
 function App() {
   return (
@@ -12,7 +13,14 @@ function App() {
           <Routes>
             <Route path="/" element={<Home />} />
             <Route path="/demo" element={<Terminal />} />
-            <Route path="/earn" element={<Earn />} />
+            <Route
+              path="/earn"
+              element={
+                <ActivityLogProvider>
+                  <EarnWithPrivyServerWallet />
+                </ActivityLogProvider>
+              }
+            />
           </Routes>
         </div>
       </Router>
